# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Recall.ai API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type UserPlan = (typeof UserPlan)[keyof typeof UserPlan];

export const UserPlan = {
  free: "free",
  pro: "pro",
} as const;

export interface User {
  id: number;
  email: string;
  /** @nullable */
  username?: string | null;
  plan: UserPlan;
  monthlySavesCount: number;
  savesLimit: number;
  preferredLanguage: string;
  onboardingCompleted: boolean;
  createdAt: string;
}

export interface RegisterBody {
  email: string;
  password: string;
  /** @nullable */
  username?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateProfileBody {
  /** @nullable */
  username?: string | null;
  /** @nullable */
  preferredLanguage?: string | null;
  /** @nullable */
  onboardingCompleted?: boolean | null;
}

export interface SummarizeBody {
  /** @nullable */
  url?: string | null;
  /** @nullable */
  preferredLanguage?: string | null;
}

export type SummaryResultSourceType =
  (typeof SummaryResultSourceType)[keyof typeof SummaryResultSourceType];

export const SummaryResultSourceType = {
  url: "url",
  youtube: "youtube",
  pdf: "pdf",
  docx: "docx",
  txt: "txt",
} as const;

export interface SummaryResult {
  title: string;
  verdict: string;
  bullets: string[];
  /** @nullable */
  articleText?: string | null;
  language: string;
  recallScore: number;
  credibilityScore: number;
  /** @nullable */
  credibilityVerdict?: string | null;
  /** @nullable */
  url?: string | null;
  sourceType: SummaryResultSourceType;
}

export interface SuggestQuestionsBody {
  title: string;
  verdict: string;
  bullets: string[];
}

export interface SuggestQuestionsResponse {
  questions: string[];
}

export interface AskAboutContentBody {
  question: string;
  title: string;
  verdict: string;
  bullets: string[];
  /** @nullable */
  articleText?: string | null;
}

export interface AskLibraryBody {
  question: string;
}

export type AskResponseCitationsItem = {
  title: string;
  /** @nullable */
  id?: number | null;
};

export interface AskResponse {
  answer: string;
  citations: AskResponseCitationsItem[];
}

export interface RabbitHoleBody {
  title: string;
  verdict: string;
  bullets: string[];
}

export type RabbitHoleResponseSuggestionsItem = {
  title: string;
  description: string;
  searchQuery: string;
};

export interface RabbitHoleResponse {
  suggestions: RabbitHoleResponseSuggestionsItem[];
}

export interface SavedArticle {
  id: number;
  userId: number;
  /** @nullable */
  url?: string | null;
  title: string;
  verdict: string;
  bullets: string[];
  /** @nullable */
  articleText?: string | null;
  language: string;
  recallScore: number;
  credibilityScore: number;
  /** @nullable */
  credibilityVerdict?: string | null;
  /** @nullable */
  collectionId?: number | null;
  /** @nullable */
  shareToken?: string | null;
  isPublic: boolean;
  /** @nullable */
  lastReviewedAt?: string | null;
  reviewCount: number;
  sourceType: string;
  createdAt: string;
}

export interface SavedArticlesResponse {
  articles: SavedArticle[];
  total: number;
  page: number;
  totalPages: number;
}

export interface SaveArticleBody {
  /** @nullable */
  url?: string | null;
  title: string;
  verdict: string;
  bullets: string[];
  /** @nullable */
  articleText?: string | null;
  language: string;
  recallScore: number;
  credibilityScore: number;
  /** @nullable */
  credibilityVerdict?: string | null;
  /** @nullable */
  collectionId?: number | null;
  sourceType: string;
}

export interface UpdateCollectionBody {
  /** @nullable */
  collectionId?: number | null;
}

export interface ShareResponse {
  shareToken: string;
  shareUrl: string;
}

export type LibraryStatsTopTopicsItem = {
  topic: string;
  count: number;
};

export interface LibraryStats {
  totalSaves: number;
  monthlySaves: number;
  savesLimit: number;
  avgRecallScore: number;
  avgCredibilityScore: number;
  topTopics: LibraryStatsTopTopicsItem[];
  recentSaves: SavedArticle[];
}

export interface Collection {
  id: number;
  userId: number;
  name: string;
  color: string;
  articleCount: number;
  createdAt: string;
}

export interface CreateCollectionBody {
  name: string;
  color: string;
}

export interface Highlight {
  id: number;
  userId: number;
  articleId: number;
  bulletText: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  articleTitle?: string | null;
  createdAt: string;
}

export interface CreateHighlightBody {
  articleId: number;
  bulletText: string;
  /** @nullable */
  note?: string | null;
}

export interface TopicBreakdown {
  topic: string;
  count: number;
  percentage: number;
}

export interface ReadingDna {
  topicBreakdown: TopicBreakdown[];
  totalArticles: number;
  avgSessionLength: number;
  mostActiveDay: string;
  preferredContentType: string;
}

export interface ReadingStreak {
  currentStreak: number;
  longestStreak: number;
  totalDaysRead: number;
  /** @nullable */
  lastReadAt?: string | null;
}

export type MentorRecommendationsKnowledgeGapsItem = {
  topic: string;
  description: string;
  suggestedReads: string[];
};

export interface MentorRecommendations {
  knowledgeGaps: MentorRecommendationsKnowledgeGapsItem[];
  insights: string;
}

export interface CompleteReviewBody {
  remembered: boolean;
}

export interface DigestPreview {
  articles: SavedArticle[];
  generatedAt: string;
}

export interface WeeklyReport {
  weekStart: string;
  weekEnd: string;
  totalSaves: number;
  topThemes: string[];
  topSaves: SavedArticle[];
  insights: string;
}

export interface PaymentOrder {
  orderId: string;
  amount: number;
  currency: string;
  keyId: string;
}

export interface VerifyPaymentBody {
  orderId: string;
  paymentId: string;
  signature: string;
}

export type GetSavedArticlesParams = {
  page?: number;
  limit?: number;
  /**
   * @nullable
   */
  collection_id?: number | null;
  /**
   * @nullable
   */
  search?: string | null;
  sort?: GetSavedArticlesSort;
};

export type GetSavedArticlesSort =
  (typeof GetSavedArticlesSort)[keyof typeof GetSavedArticlesSort];

export const GetSavedArticlesSort = {
  created_at: "created_at",
  recall_score: "recall_score",
  credibility_score: "credibility_score",
  title: "title",
} as const;