# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Recall.ai API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /collections:
    get:
      operationId: getCollections
      tags: [collections]
      summary: List user collections
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
    post:
      operationId: createCollection
      tags: [collections]
      summary: Create a new collection
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollectionBody"
      responses:
        "201":
          description: Collection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
  /collections/{id}:
    delete:
      operationId: deleteCollection
      tags: [collections]
      summary: Delete a collection
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /highlights:
    get:
      operationId: getHighlights
      tags: [highlights]
      summary: List user highlights
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of highlights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Highlight"
    post:
      operationId: createHighlight
      tags: [highlights]
      summary: Save a bullet as a highlight
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHighlightBody"
      responses:
        "201":
          description: Highlight saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Highlight"
  /highlights/{id}:
    delete:
      operationId: deleteHighlight
      tags: [highlights]
      summary: Delete a highlight
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: ["string", "null"]
        plan:
          type: string
          enum: [free, pro]
        monthlySavesCount:
          type: integer
        savesLimit:
          type: integer
        preferredLanguage:
          type: string
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - plan
        - monthlySavesCount
        - savesLimit
        - preferredLanguage
        - onboardingCompleted
        - createdAt
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        username:
          type: ["string", "null"]
      required:
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    UpdateProfileBody:
      type: object
      properties:
        username:
          type: ["string", "null"]
        preferredLanguage:
          type: ["string", "null"]
        onboardingCompleted:
          type: ["boolean", "null"]
...
    Collection:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        color:
          type: string
        articleCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - color
        - articleCount
        - createdAt
    CreateCollectionBody:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
      required:
        - name
        - color
    Highlight:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        articleId:
          type: integer
        bulletText:
          type: string
        note:
          type: ["string", "null"]
        articleTitle:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - articleId
        - bulletText
        - createdAt
    CreateHighlightBody:
      type: object
      properties:
        articleId:
          type: integer
        bulletText:
          type: string
        note:
          type: ["string", "null"]
      required:
        - articleId
        - bulletText
...