# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Recall.ai API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserPlan } from "./userPlan";

export interface User {
  id: number;
  email: string;
  /** @nullable */
  username?: string | null;
  plan: UserPlan;
  monthlySavesCount: number;
  savesLimit: number;
  preferredLanguage: string;
  onboardingCompleted: boolean;
  createdAt: Date;
}